+++ web/src/routes/admin/library-management/[id]/edit/+page.ts
@@ -1,18 +1,15 @@
-import { authenticate, requestServerInfo } from '$lib/utils/auth';
+import { authenticate } from '$lib/utils/auth';
 import { getFormatter } from '$lib/utils/i18n';
-import { searchUsersAdmin } from '@immich/sdk';
 import type { PageLoad } from './$types';
 
 export const load = (async ({ url }) => {
   await authenticate(url, { admin: true });
-  await requestServerInfo();
-  const allUsers = await searchUsersAdmin({ withDeleted: true });
+
   const $t = await getFormatter();
 
   return {
-    allUsers,
     meta: {
-      title: $t('admin.user_management'),
+      title: $t('admin.library_details'),
     },
   };
 }) satisfies PageLoad;
