+++ web/src/routes/admin/users/(list)/new/+page.ts
@@ -4,12 +4,11 @@ import type { PageLoad } from './$types';
 
 export const load = (async ({ url }) => {
   await authenticate(url, { admin: true });
-
   const $t = await getFormatter();
 
   return {
     meta: {
-      title: $t('admin.library_details'),
+      title: $t('admin.user_management'),
     },
   };
 }) satisfies PageLoad;
